[DBC] float-based item budget values initial implementation (#5799)
* [DBC] float-based item budget values initial implementation

* apply rounding only for ptr

* remove unnecessary cast

* use float values for spelldata values tied to item budget

* use float damage_replace_stat and damage_secondary

* use double instead of float

* use float value for demonhunter & druid calculate_expected_max_health()

* Revert "use float damage_replace_stat and damage_secondary"

This reverts commit ef30b0695ee3c43f60ccd9177bd9c79fdfcebfbd.

* Revert "use float values for spelldata values tied to item budget"

This reverts commit fe020abe1a271de2a4313949c91a82f2b898afac.

* Revert "use float value for demonhunter & druid calculate_expected_max_health()"

This reverts commit 5a4f630b0ec37b65f7f249ca781c3c444aba279d.

* remove doubling of fields

* change format json field names

* remove non-relevant .inc files

+++ engine/dbc/rand_prop_points.hpp
@@ -12,8 +12,8 @@
 struct random_prop_data_t
 {
   unsigned ilevel;
-  unsigned damage_replace_stat;
-  unsigned damage_secondary;
+  double   damage_replace_stat;
+  double   damage_secondary;
   float    p_epic[5];
   float    p_rare[5];
   float    p_uncommon[5];
